refactor: add model package
