+++ backend/internal/repositories/post_repository.go
@@ -2,28 +2,10 @@ package repositories
 
 import (
 	"myapp/internal/entities"
+	"myapp/internal/repositories/model"
 
 	"gorm.io/gorm"
 )
-
-type Post struct {
-	Id        int64  `gorm:"column:id"`
-	Title     string `gorm:"column:title"`
-	Body      string `gorm:"column:body"`
-	UserId    int64  `gorm:"column:user_id"`
-	CreatedAt string `gorm:"column:created_at"`
-	UpdatedAt string `gorm:"column:updated_at"`
-}
-
-type PostWithUsername struct {
-	Id        int64  `gorm:"column:id"`
-	Title     string `gorm:"column:title"`
-	Body      string `gorm:"column:body"`
-	UserId    int64  `gorm:"column:user_id"`
-	Username  string `gorm:"column:username"`
-	CreatedAt string `gorm:"column:created_at"`
-	UpdatedAt string `gorm:"column:updated_at"`
-}
 
 type PostRepository struct {
 	Conn *gorm.DB
@@ -36,7 +18,7 @@ func NewPostRepository(conn *gorm.DB) *PostRepository {
 }
 
 func (r *PostRepository) GetAll(limit, offset int64) ([]*entities.Post, error) {
-	var posts []*PostWithUsername
+	var posts []*model.PostWithUsername
 	if err := r.Conn.Table("posts").Select("posts.*, users.name as username").Joins("JOIN users ON posts.user_id = users.id").
 		Order("posts.id").Limit(int(limit)).Offset(int(offset)).Scan(&posts).Error; err != nil {
 		return nil, err
@@ -44,14 +26,14 @@ func (r *PostRepository) GetAll(limit, offset int64) ([]*entities.Post, error) {
 
 	var result []*entities.Post
 	for _, post := range posts {
-		result = append(result, convertPostWithUsernameToEntity(post))
+		result = append(result, model.ConvertPostWithUsernameToEntity(post))
 	}
 
 	return result, nil
 }
 
 func (r *PostRepository) GetById(postId int64) (*entities.Post, error) {
-	var post PostWithUsername
+	var post model.PostWithUsername
 	if err := r.Conn.Table("posts").Select("posts.*, users.name as username").Joins("JOIN users ON posts.user_id = users.id").
 		Where("posts.id = ?", postId).Scan(&post).Error; err != nil {
 		return nil, err
@@ -60,21 +42,9 @@ func (r *PostRepository) GetById(postId int64) (*entities.Post, error) {
 	if len(post.Username) == 0 {
 		return nil, nil
 	}
-	return convertPostWithUsernameToEntity(&post), nil
+	return model.ConvertPostWithUsernameToEntity(&post), nil
 }
 
 func (r *PostRepository) Delete(postId int64) error {
-	return r.Conn.Delete(&Post{}, postId).Error
-}
-
-func convertPostWithUsernameToEntity(v *PostWithUsername) *entities.Post {
-	return &entities.Post{
-		Id:        v.Id,
-		Title:     v.Title,
-		Body:      v.Body,
-		UserId:    v.UserId,
-		Username:  v.Username,
-		CreatedAt: v.CreatedAt,
-		UpdatedAt: v.UpdatedAt,
-	}
+	return r.Conn.Delete(&model.Post{}, postId).Error
 }
